refactor(merchant acquirer): edit calculate fee logic and generate amount

+++ services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/domain/entity/Merchant.java
@@ -23,5 +23,5 @@ public class Merchant {
   private String acquirerId;
   @Column(name = "acquiring_fee", precision = 5, scale = 4)
   private BigDecimal acquiringFee;
-  private int averageCheck;
+  private BigDecimal averageCheck;
 }
